Increase no_sprite player's size
When the AddAnimatedSprite feature is added, the player is resized.
If the player is standing on the ground when this happens, they get
stuck in the ground because the animated sprite's size is slightly
taller than the no_sprite's size. So the no_sprite's size was increased
to have the same height as the animated sprite's size. (hacky but works)