Add properties 'always_loaded', 'loader', 'solid' to all enemies
To reduce lag, not all enemies need to be solid or be loaders.
Only Ground enemies need to be solid.
Any ground enemies that need to be always_loaded should also be loaders.
Flying enemies that are always_loaded do not need to be loaders,
because they don't touch any solids.

TODO: Set properties for all enemies